Change return type of getpagesize() to size_t

+++ include/cat_memory.h
@@ -95,7 +95,7 @@ CAT_API void *cat_realloc_function(void *ptr, size_t size);
 CAT_API void cat_free_function(void *ptr);
 CAT_API void cat_freep_function(void *ptr); /* free(ptr->ptr) */
 
-CAT_API int cat_getpagesize(void);
+CAT_API size_t cat_getpagesize(void);
 CAT_API void *cat_getpageof(const void *ptr);
 
 CAT_API unsigned int cat_bit_count(uintmax_t num);
